% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grandpa.R
\name{addAugment}
\alias{addAugment}
\title{Add degree augmentation label}
\usage{
addAugment(G, nbins, degType)
}
\arguments{
\item{G}{an igraph object.}

\item{nbins}{the number of degree bins.}

\item{degType}{string; type of degree (see ?degree).}
}
\value{
an igraph object with `augLabel` vertex attribute.
}
\description{
Add degree augmentation label
}
