% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grandpa.R
\name{addCommunityAugment}
\alias{addCommunityAugment}
\title{Add community augmentation label}
\usage{
addCommunityAugment(G, resolution = 1, seed = NULL)
}
\arguments{
\item{G}{an igraph object.}

\item{resolution}{Louvain resolution parameter (default 1).}

\item{seed}{random seed for reproducibility (default NULL = no seed).}
}
\value{
an igraph object with `CommunityLabel` vertex attribute.
}
\description{
This function is typically called by the `grandpa` function rather than
directly. By default it implements Louvain clustering to detect
community membership. For best results we recommend fitting a community
algorithm outside of GRANDPA and attaching the result as the
`CommunityLabel` vertex attribute prior to calling `grandpa()`.
}
