% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grandpa.R
\name{calcObsProbs}
\alias{calcObsProbs}
\title{Calculates the observed edge and vertex probabilities}
\usage{
calcObsProbs(G)
}
\arguments{
\item{G}{An igraph object containing vertex attributes whose names
include the substring 'Label' (case-insensitive).}
}
\value{
a list of tidy tables containing the vertex probabilities and
  edge probabilities respectively.
}
\description{
This function is typically called by the grandpa function rather than
directly.
}
