% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grandpa.R
\name{createMap}
\alias{createMap}
\title{Creates the map of label combinations to vertex IDs}
\usage{
createMap(sampled_v)
}
\arguments{
\item{sampled_v}{data frame of sampled vertices with an `ix` column.}
}
\value{
a list mapping label combinations to integer vertex IDs.
}
\description{
Creates the map of label combinations to vertex IDs
}
