% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grandpa.R
\name{degree_error}
\alias{degree_error}
\title{Calculates the NRMSE between the degree distributions of two graphs}
\usage{
degree_error(G, Gt)
}
\arguments{
\item{G}{the original graph.}

\item{Gt}{the generated graph.}
}
\value{
numeric; the NRMSE between the degree distributions.
}
\description{
Calculates the NRMSE between the degree distributions of two graphs
}
