% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grandpa.R
\name{grandpa}
\alias{grandpa}
\title{Conducts a GRANDPA procedure}
\usage{
grandpa(
  G,
  nt = vcount(G),
  mt = ecount(G),
  preventSelf = TRUE,
  preventDups = TRUE,
  augment = FALSE,
  augmentCommunity = FALSE,
  nbins = 3,
  degType = "all",
  seed = 0
)
}
\arguments{
\item{G}{an igraph object containing the original graph. Attributes that
will be used in the modeling procedure must contain the word 'Label'
(case-insensitive) in their attribute name.}

\item{nt}{the number of desired nodes in the generated graph.}

\item{mt}{the number of desired edges in the generated graph.}

\item{preventSelf}{logical; should self connections be prevented?}

\item{preventDups}{logical; should duplicate connections be prevented?}

\item{augment}{logical; should degree augmentation occur?}

\item{augmentCommunity}{logical; should communities be augmented?}

\item{nbins}{the number of degree bins used in the degree augmentation.}

\item{degType}{string; the type of degree to be used in the degree
augmentation. See ?degree for details.}

\item{seed}{a random seed set for reproducibility.}
}
\value{
an igraph object with attribute and augmentation labels.
}
\description{
Uses the GRANDPA framework to generate a network with matching attribute
probabilities.
}
