% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grandpa.R
\name{initializeTargetG}
\alias{initializeTargetG}
\title{Initializes an empty graph with the generated vertices}
\usage{
initializeTargetG(sampled_v, nt, d)
}
\arguments{
\item{sampled_v}{data frame of sampled vertices and their labels.}

\item{nt}{number of desired nodes.}

\item{d}{logical; is the graph directed?}
}
\value{
an empty igraph object with the desired vertices.
}
\description{
Initializes an empty graph with the generated vertices
}
