% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/physNet_data.R
\docType{data}
\name{physNet}
\alias{physNet}
\title{Simulated physician network data}
\format{
An object of class \code{"igraph"}; see \code{\link[igraph]{igraph}}.
}
\source{
\href{https://arxiv.org/abs/2211.15000}{arXiv}
}
\usage{
data(physNet)
}
\description{
Graph simulated from medicare claims wherein
physicians are nodes, and edges between nodes indicate overlapping patients.
Physician specialty is represented as a vertex attribute and shown using vertex
color by default.
}
\examples{
data(physNet)
plot(physNet,vertex.label=NA)
}
\references{
Bobak et al. (2022) arXiv preprint arXiv:2211.15000.
(\href{https://arxiv.org/abs/2211.15000}{arXiv})
}
\keyword{datasets}
