% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grandpa.R
\name{preventDuplicateConnections}
\alias{preventDuplicateConnections}
\title{Prevents duplicate connections during graph generation}
\usage{
preventDuplicateConnections(v1, v2, lab1, lab2, C2V, Gt)
}
\arguments{
\item{v1}{a vertex id.}

\item{v2}{a second vertex id.}

\item{lab1}{the label corresponding to vertex id 1.}

\item{lab2}{the label corresponding to vertex id 2.}

\item{C2V}{mapping of label categories to vertex IDs.}

\item{Gt}{the graph being generated.}
}
\value{
updated vertex IDs.
}
\description{
Prevents duplicate connections during graph generation
}
