% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grandpa.R
\name{sampleEdges}
\alias{sampleEdges}
\title{Creates the edges that will be used in the generated graph}
\usage{
sampleEdges(observedEdgeLabels, mt, seed)
}
\arguments{
\item{observedEdgeLabels}{data frame of edge label combinations and
their probabilities.}

\item{mt}{number of desired edges.}

\item{seed}{random seed.}
}
\value{
a data frame of sampled edges.
}
\description{
Creates the edges that will be used in the generated graph
}
