% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grandpa.R
\name{sampleVertices}
\alias{sampleVertices}
\title{Creates the vertices that will be used in the generated graph}
\usage{
sampleVertices(observedLabels, nt, seed)
}
\arguments{
\item{observedLabels}{a data frame containing observed label
combinations and their probabilities from the original graph.}

\item{nt}{the number of desired nodes in the generated graph.}

\item{seed}{the random seed used for reproducibility.}
}
\value{
a data frame containing node attributes.
}
\description{
Creates the vertices that will be used in the generated graph
}
