% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grandpa.R
\name{simGraph}
\alias{simGraph}
\title{Procedure to add edges to the generated graph}
\usage{
simGraph(Gt, sampled_e, C2V, seed, preventSelf, preventDups)
}
\arguments{
\item{Gt}{the empty graph.}

\item{sampled_e}{data frame of sampled edges with label columns.}

\item{C2V}{mapping of label combinations to vertex IDs.}

\item{seed}{random seed.}

\item{preventSelf}{logical; prevent self connections?}

\item{preventDups}{logical; prevent duplicate connections?}
}
\value{
a generated igraph object.
}
\description{
Procedure to add edges to the generated graph
}
