% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grandpa.R
\name{addAugment}
\alias{addAugment}
\title{Add degree augmentation label}
\usage{
addAugment(G, nbins, degType)
}
\arguments{
\item{G}{an igraph object}

\item{nbins}{the number of degree bins used in the degree augmentation}

\item{degType}{string; the type of degree to be used in the degree augmentation. See ?degree for details}
}
\value{
an igraph object containing an augmentation label corresponding to degree
}
\description{
This function is typically called by the grandpa function rather than directly.
}
\examples{

G <- make_graph("Zachary")
G <- addAugment(G,3,"all")
}
