% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grandpa.R
\name{addCommunityAugment}
\alias{addCommunityAugment}
\title{Add community augmentation label}
\usage{
addCommunityAugment(G)
}
\arguments{
\item{G}{an igraph object}
}
\value{
an igraph object containing an augmentation label corresponding to community membership
}
\description{
This function is typically called by the grandpa function rather than directly. By default, this function implements louvain clustering to detect community membership. Note, it is recommended to fit a community algorithm outside of the GRANDPA framework and add a vertext label called 'CommunityLabel' for best results.
}
\examples{

G <- make_graph("Zachary")
G <- addCommunityAugment(G)
}
