% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grandpa.R
\name{degree_error}
\alias{degree_error}
\title{Calculates the error between two degree distributions of graphs}
\usage{
degree_error(Gs, Gt)
}
\arguments{
\item{Gs}{the original graph}

\item{Gt}{the generated graph}
}
\value{
numerical; the NRMSE between the degree distributions
}
\description{
This function takes an original graph and a simulated graph and calculates the normalized root mean square error between their degree distributions.
}
