% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grandpa.R
\name{grandpa}
\alias{grandpa}
\title{Conducts a GRANPDA procedure}
\usage{
grandpa(
  G,
  nt = vcount(G),
  mt = ecount(G),
  preventSelf = T,
  preventDups = T,
  augment = F,
  augmentCommunity = F,
  nbins = 3,
  degType = "all",
  seed = 0
)
}
\arguments{
\item{G}{an igraph object containing the original graph. Attributes which will be in the modelling procedure must contain the word 'Label'}

\item{nt}{the number of desired nodes in the generated graph}

\item{mt}{the number of desired edges in the generated graph}

\item{preventSelf}{logical; should self connections be prevented?}

\item{preventDups}{logical; should duplicate connections be prevented?}

\item{augment}{logical; should degree augmentation occur?}

\item{augmentCommunity}{logical; should communited be augmented?}

\item{nbins}{the number of degree bins used in the degree augmentation}

\item{degType}{string; the type of degree to be used in the degree augmentation. See ?degree for details}

\item{seed}{a random seed set for reproducibility}
}
\value{
an igraph object with attribute and augmentation labels
}
\description{
Uses the GRANPDA framework to generate a network with matching attribute probabilities
}
\examples{

G = make_graph("Zachary")
V(G)$degree<-degree(G,mode="all")

V(G)$Label1<-rep("A",vcount(G))
V(G)$Label1[V(G)$degree>median(V(G)$degree)]<-"B"

GSim<-grandpa(G)

}
